refactor(request): extract routing lock marker helpers
